Apply translate to Line
