+++ shoes-core/lib/shoes/line.rb
@@ -3,10 +3,11 @@ require 'matrix'
 class Shoes
   class Line
     include Common::UIElement
-    include Common::Stroke
-    include Common::Style
     include Common::Clickable
     include Common::Hover
+    include Common::Stroke
+    include Common::Style
+    include Common::Translate
 
     attr_reader :point_a, :point_b
